resultify: pr-fixes, move resultify to main folder
